fixed things up
