+++ src/main/java/com/zipcodewilmington/scientificcalculator/Factorial.java
@@ -1,0 +1,13 @@
+package com.zipcodewilmington.scientificcalculator;
+
+public class Factorial {
+
+    public long Factorial(int n){
+        long fact = 1;
+        for(int i =2 ; i <= n ; i++){
+            fact = fact * i;
+        }
+        return fact;
+    }
+
+}
